requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
passlib
bcrypt